requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
PyMySQL==1.1.1
pyparsing==3.2.3
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
RPi.GPIO==0.7.1
six==1.17.0
